refactor project structure and add cover image to readme
