--- llm/transcript.js
@@ -1,3 +1,0 @@
-export const TRANSCRIPT = `
-[00:00] - 
-`
