--- urls.md
@@ -1,2 +1,0 @@
-https://www.youtube.com/watch?v=jKB0EltG9Jo
-https://www.youtube.com/watch?v=RS-FuJ7rUsE
